+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.23] - 2023-11-20
+
+### Added
+
+- Deferred fluid storage
+
 ## [0.4.22] - 2023-07-26
 
 ### Fixed